Improve UX of instancesSelectEditor display value
When the menu is open, don't show the clear icon. Before this change
the UX was confusing since the user might think it would allow them to
clear the search text, which it does, but it also clears the existing
selection, which they might not want to do.

When the menu closes, clear the search text; previously, the remaining
search text could wrongly be interpreted as a selected value, which
could confuse the user.