Upgrade shacl-tulip version after fixing multi-save enum issue
Previously, if an existing blank node was edited and saved again
shacl-tulip 0.0.5 would not correctly delete all existing related
quads first before saving them again. This caused an issue of changed
enum selections (shacl:in) to be saved as an rdf array in the quad
store. The shacl-tulip fix now correctly determines the node type
in order to find the correct existing matching quads and then delete
them before regenerating the latest quads related to the saved node.
This causes the enum selection to be saved as the single latest
selection