Update 'display_name_autogenerate' operation and handling
These changes are made in response to the need for viewing and filtering by
by the display labels of related records and properties of a given record,
both in the 'InstancesSelectEditor' and in the 'NodeShapeViewer' on the main app
view. The absence of this feature was noticed when records on the main view
rendered all their properties seemingly as pids, while a 'display_name_autogenerate'
option was in fact provided for the class types of those properties.

The main change was made to the 'utils' module to introduce a new 'getRecordDisplayLabel'
function, which is used by the 'NodeShapeViewer' to find the relevant display label
for each property in the record. This label gets passed as a new 'displayLabel' prop
to the 'NamedNodeViewer', which will render it if available and if prioritized.
The priority order for rendering is taken as prefLabel > displayLabel > pid. In
addition the 'getRecordDisplayLabel' will also recurse in order to find the display
label of related records, if the property associated with a related record is
included in the current record's display label.

For example, if we have a 'display_name_autogenerate' option:
dldemoed:XYZSubject: '{dldemoed:name} ({dldemoed:subject_type})', where 'dldemoed:name'
is a literal property of a 'XYZSubject' record, and 'dldemoed:subject_type' is the pid
of a related 'XYZSubjectType' record, then it will try to find the 'display_name_autogenerate'
option for 'XYZSubjectType' too, and generate it for the related record, and then
in turn use it for the display label of the current record.

Another change that was necessary to support this recursion was to fix the
behavior of the 'service_fetch_before' option. If a 'display_name_autogenerate'
option requires recursion, it also requires the related record to be available
locally. That won't be the case unless that record, or records of that class,
have been fetched before. So a quick fix is to fetch them at startup if they are
only a handful. The 'service_fetch_before' endpoint option for 'get-records' was
changed internally to 'get-paginated-records' so that it actually works with the
latest dumpthings integration.

Lastly, since the display label generation now works for records on the main app
view and in the 'InstancesSelectEditor', the filtering behavior of these components
were updated so that they filter also on the display label. For all cases, the filtering
searches for matches with preflabel or display label or pid.