Allow constrained and user-controlled display of amount of related records
This pertains to both the 'NodeShapeViewer', which displays the literal nodes,
named nodes and blank nodes of a record being viewed, as well as the
'PropertyShapeEditor', which renders all form fields of a record being edited.
It is possible that some related nodes of a record will have a list of multiple
values. For example, a distribution might have many parts and it is bad UX if
e.g. 500 (or a million) parts are rendered automatically for a given distribution
record, both in the record viewer and record editor components.

For this reason, a new feature was introduced whereby a default number of records
will be rendered automatically, and the user will then be able to decide whether
to show more records, or to show fewer, by using dedicated buttons.

This was achieved as follows:
- new config options were introduced, one per viewer/editor:
  - 'editor_config[PropertyShapeEditor][recordNumberStepSize]' with a default of 5
  - 'viewer_config[NodeShapeViewer][recordNumberStepSize]' with a default of 5
- the config options define how many records to show by default, and the user can
  increase this to the max corresponding to the number of records, and to the min
  corresponding to the default stepsize
- the new UX code was added explicitly to the 'PropertyShapeEditor', and then via the
  new 'MoreOrLessRecordsViewer' to the 'NodeShapeViewer'

For this new functionality to work seamlessly with the 'PropertyShapeEditor's
ability to add/remove triples for a given form property, the instantiation of any
given editor component (via the 'component' tag) had to be updated. This was
necessary because the previous state did not use the 'key' of the component
correctly: it included the triple index where it should include the triple itself.
This is because the triple's position in the data array might change, but the
triple node itself won't change. So using the index in the 'key' actually lead to
unnecessary new components being instantiated every time the 'key' changed; and
this also lead to bugs in triple display names which was how the issue was
noticed in the first place. After figuring this out and using the triple node
itself in the key, the bug was fixed. In addition, addition of a triple to the
data was updated to use splice in order to add the triple at the current index,
and not just to push to the last position in the array.