Update shacl-tulip dependency, use unique component-keys to prevent re-rendering
The 'key' field for the 'component' tag (that renders editor components dynamically)
was changed in d8a933f2c52e1cc8e87295f8180a103a11f95d5f to use the current triple
object value. The original motivation for changing the key was to prevent unnecessary
re-rendering in the case of multiple values for a property (which came up when working
on limiting the number of values that are shown, useful in cases where there are
hundreds/thousands/millions). E.g. if we have a property 'keywords' and we have 7 of
them, deleting one would cause all others with an array position after the deleted one
to be re-rendered, because their key changed (because of the position change). That is
why using the index in the key (the original state) was not useful.

Changing it to the triple object value, i.e. the array element, seemed to work fine for
an editor such as the 'InstancesSelectEditor', because that component value changes on
selection of an item, once. But for a text field editor the problem became obvious:
the field's focus shifted away every single time a single character was entered into
the field, because of re-rendering.

This provided the motivation to update shacl-tulip to allow adding additional fields to
a triple object value.

In this current commit, we bump the shacl-tulip version to that supporting state.
Additionally, we change all parts of the shacl-vue code that accessed a triple object
value in some way, so that it now accesses the 'value' field of the javascript object.
The important change is also to generate a random unique key per triple object that
shacl-vue renders, which is stored as the '_key' field in the javascript object now
used by shacl-tulip. This '_key' is then used as the component-key to prevent
unnecessary rerendering.