PID query parameter and exact vs partial matching
This commit adds the ability to filter by PID via query parameter, and associated
functionality. Before this, a pid could be included as a query parameter but only
as a fully resolved IRI and it only served to highlight the identifed record. This
was problematic in long lists, where the user would have to scroll to wherever they
found the highlighted record.

A PID can now be included as a URL query parameter, either as CURIE or IRI, using
'pid=' and then the url-encoded identifier. This will populate the 'searchText' field
with the resolved IRI and use exact matching to find the record (see below). In
addition, if the '&edit=true' query parameter is also included, a form to edit that
specific record will be opened.

URL updates are done to fit in with the existing handling. Wether a record is found
for the provided PID or not, the pid query parameter is removed from the URL once
processed, unless '&edit=true' is also included, in which case both parameters
remain part of the URL.

To support the case where the exact record needs to be filtered to, as opposed to
also showing all other records that might have the PID string of the desired record
embedded in their own PID, the string matching was updated. If the 'searchText' is
populated either via the PID query parameter, or when a user navigates internally to
a specific record, the filtering uses an exact match. If not (e.g. when a user types
in the 'searchText'), a partial approach is used. To accommodate this behavior, a
new funtion runs for every time the user enters a value in the 'searchText' field.