Fix 'NodeShapeViewer' update issue
This issue was introduced in d8a933f2c52e1cc8e87295f8180a103a11f95d5f with the
functionality to allow constrained and user-controlled display of the amount of
related records of a particular kind in the 'NodeShapeViewer'. The problem had
two parts:
- 'onUpdated' for the 'NodeShapeViewer' was not being called as expected every
  time a form is saved, which meant the record was not updated from the most
  recent state of the graph
- 'initShowCounts' is crucial to set reactively, because there are conditional
  rendering clauses in the template that depend on it; this is why we did not see
  updated values for edited fields.
The fix now uses 'watch' the react to 'lastSavedNode'. If the last saved node is
the same as the current 'NodeShapeViewer' instance, the 'updateRecord' and
'initShowCounts' functions are called and edits are visible.