Fix faulty association class selection for 'InstancesSelectEditor'
The changes introduced in c12020762401b4401f7980d478d2f736692e620f helped the
'InstancesSelectEditor' to hide duplicate association class records and to create
the necessary new blank nodes and correct linkage whenever such records are selected.
However, this also introduced an issue whereby the act of saving an association class
record did not assign it to the selected value of the originating 'InstancesSelectEditor'
as was expected to happen. The problem was that the 'lastSavedNode' 'watch'er code
did not account for the effects of the new deduplication code when it searched through
the items to find the recently saved one. This commit adds the ability to find the
recently saved blank node record via the subject IDs of its 'relatedQuads' property,
because the 'itemQuad' was assigned anew by the deduplication code, but the
'relatedQuads' were not. It then also calls the 'selectItem' function when the correct
item is found in order to trigger the correct handling of temporary blank node records.

Another minor change is to hide unnecessary item info from the item display in the
component, in order to make it easier for users to find a specific item in the list.