Update state handling between form data and graph data
Prior to this update, submitting a record happened from the latest state of a node
in 'formData', which could be different from the state that the user expected it to
be in. This occured when the 'formData' of that specific node was edited previously
in the same sesion. It was implemented as follows:

- saving a record did not remove the node from 'formData' because it is required for
  submission
- creating then saving a record put the new quads in the graph store and left the node
  data in 'formData' for future editing and submission
- editing then cancelling a previously saved record left the edited changes to the node
  in 'formData'

This lead to viewer data being out of sync with the changes the user made during editing
and to submitted data containing changes that the user did not save. Additionally, an
incorrect implementation of blank node recursion in the submission code resulted in
submitted quads not including all recursive blank node quads that the user might have
entered and saved during editing, i.e. missing data.

To fix this, several changes were made:

- A new 'getRecordQuads' utility function was implemented to correctly recurse into a
  graph from the top-level named node and to collect and resolve all blank nodes into
  the output. This is now used in the submission code as well as in the 'NodeShapeViewer'
  component.
- Submissions are not completely made from the graph store and not from 'formData'. To
  achieve this, the 'submitRdfData' function (previously 'submitFormData') was moved and
  updated from the 'useForm' composable into the 'useData' composable. It uses the
  'getRecordQuads' function to get all quads related to an edited node, instead of the
  'formNodeToQuads' from 'shacl-tulip'.
- Created/edited nodes are always removed from 'formData' when a form is either saved or
  cancelled. Also, the related logic to track the 'editMode' was simplified, since the
  'form' vs 'graph' states of this flag is not necessary anymore. The flag is now simply
  set to true if a form is opened to edit an existing record, vs false otherwise.