Fix InstanceSelectEditor's subclass derivation and item generation
When a form field had the range of a base-level class, e.g. Thing, the
'InstanceSelectEditor' did not show any items in the dropdown list and it
did not show all expected subclasses in the 'Add new item' button dropdown.
This was because the N3-store representation of the OWL graph (one of the
main inputs to a 'shacl-vue' deployment) does not contain an 'rdfs:subClassOf'
statement for each instance of a class being a subclass of another. It jumps
relationships that needed to be inferred. E.g. if 'Person' is a subclass of
'Entity' and 'Entity' is a subclass of 'Thing', the store might not have a
statement saying that 'Person' is a subclass of 'Thing', even though that is
true and, importantly, the 'InstancesSelectEditor' expected that.

To fix this a new utility function 'getSubClasses' was created to correctly
traverse the whole 'classDS' dataset and create all these interim statements
and generate an object with all classes as keys and an array of all their
respective subclasses as values. The 'InstancesSelectEditor' was updated to
use this as the source for the classes to show when the user selected the
'Add item' option, and also as the basis for making fetch calls to the backend.

It was noted that the complete list of subclasses might not all be necessary
to show to the user, and should be filtered down. The new filtering approach
uses the 'includeClass' function in the 'ShaclVue' component (via 'provide'-
'inject') along with existing config options to determine whether a class should
be included in the list of items that can be added. The logic, in order, is:
- assume we include class by default
- if class is not found in 'show_classes' or 'show_classes_with_prefix', exclude it
- if class should not be explicitly excluded, based on 'hide_classes' or
  'hide_classes_with_prefix', include it
- if the class is not included in 'no_edit_classes', include it.

This acts as the default filtering approach, and allows for a future editor-level
config approach to override the behavior.