Make ShaclORClassEditor instantiate the InstancesSelectEditor
Previously, the ShaclORClassEditor had a subset of the functionality of the InstancesSelectEditor
and it became increasingly out of sync, even while it would be ideal to retain much of the
functionality of the latter. This was not ideal. It was decided to use the same exact functionality
for both, with lead to the decision to instantiate the InstancesSelectEditor directly from the
ShaclORClassEditor. From this point onward, barring substantial API changes, the ShaclORClassEditor
functionality will always be a superset of the InstancesSelectEditor functionality.