InstancesSelectEditor: fix issue with add-new-item button doing nothing
This is because a recent commit introduced a filtering mechanism for the list of subclasses
shown when 'Add New Item' is clicked. This filtering uses the same logic as that for the
classes in the main app 'Data types' pane. However, this filtering approach does not suit
the needs of a particular InstancesSelectEditor, and a per-class config approach would be
preferrable. This is deferred to future developments, and in the mean time a temporary fix
is applied: if the list of subclasses only includes the class itself, then render that class
instead of applying the filtering approach (which would likely result in zero elements in the
list). On second thought, this seems like a useful check to have, not only temporary.