Alpha version of support for uploading a file via 'git-annex-p2phttp'
This introduces the first (incomplete) version of the components and functionality necessary
to support file upload to a 'git-annex-p2phttp' server, which is supported by 'forgejo'. The
main component is the 'GitAnnexUploader':
- provides a drag-and-drop/click-and-select upload button as UI
- determines file properties from the javascript file object
- calculates a 'SHA-256' hash of the uploaded file
- constructs git annex key:
- grabs credentials from shacl-vue token management composable
- constructs the upload url and makes the POST request

A new config option 'gitannex_p2phttp_config' supports the key and url construction:
- 'base_url': e.g. 'https://trr379.de/git-annex-p2phttp/git-annex' (should be set to
  '/forgejo-api', along with a vite config update) during local development for CORS reasons
- 'annex_uuid': the UUID of the git-annex repository where the files will be uploaded to
- 'client_uuid': arbitrarily set to 'shacl-vue'

The other main component is the 'DistributionUploadEditor', which is responsible for instantiating
the 'GitAnnexUploader' based on matching criteria. The current matching criteria is that the
schema slot should have an annotation 'shaclvue:gitAnnexUpload: true', with 'shaclvue' being
a new namespace: 'https://concepts.datalad.org/ns/shaclvue/'. The component will then assume
that the range is a 'Distribution', more specifically 'trr379ra:TRR379Distribution', and will
also instantiate an 'InstancesSelectEditor' as it usually would for a slot without the annotation.
This is of course not generalized and will only work for that specific schema version and 'shacl-vue'
deployment. TODO: this aspect will be changed in future, to allow higher level generalization
that assumes no specific schema or UI deployment version.

On successful upload, the 'GitAnnexUploader' component emits a 'uploadComplete' event to its
parent 'DistributionUploadEditor' component, which includes all the file and upload details.
These are used to create the necessary quads that describes the uploaded file and its relation
to the parent record. For this, the 'id_prefix' config option is necessary and also provided
via the 'gitannex_p2phttp_config' object; it is prepended to the uploaded file's git-annex key
in order to create the PID for the newly created 'Distribution' record. Other created quads:
- blank node for the checksum
- blank node for the statement that adds the download url as 'dcat:downloadUrl'
- byte size as a literal property of the 'Distribution'
- file name is set via the 'prefLabel' slot of the 'Distribution'

Several changes to other components were necessary to support all of the above:
- Add 'isInternalUpdate' to the 'base' composable, which allows distinction between an update
  to component model value that is set internally, or that is set from its parent. This was
  required to allow the 'InstancesSelectEditor' to rerender its selected item display value
  once the 'DistributionUploadEditor' (i.e. its parent) sets the model value. Accordingly,
  'InstancesSelectEditor' is updated with a new watcher that will ignore internal updates
  and for external updates it will set the 'isInternalUpdate' flag to false again.
- A new 'tokenWarning' ref is provided by the 'ShaclVue' component and can be injected into
  any components as a means to provide a standard API to trigger a token warning. This is used
  'GitAnnexUploader', when a user interacts with the upload UI, a check is first done to see if
  a token is set, and if not, it will trigger a token warning. The 'AppHeader' component has new
  code to react to the token warning, which it does by making the token icon flash for several
  seconds.