Introduce generic handling of component-specific config and text content
The idea here is that any component-specific configuration options
can be provided and any component should be able to load these options
if provided. To achieve this, a new composable 'useCompConfig' has been
added. It depends on a change to the editor loading module, where an
explicit name is defined for the component based on its filename upon
load. This name is retrieved by the composable for any specific component
and the associated config option is then loaded for that component. This
approach is also used in the new viewer loading module.

Another necessary change was to update the configuration composable to
work with generic 'component_config' to specify these options rather than
editor or viewer component config as was done previously.

Several components were updated to make use of this generic options
loading approach.

Another new configuration option is added into this commit: allowing
for generic text content loading. The new config option 'content' allows
users to specify source URLs for any number of text documents, either
served together with the application, or anywhere accessible online. These
will be prefetched by the application, using the new 'loadAllContent' function
and made available to components via config. Other config options can point
to such content by prepending 'content:' to the key in the 'content' option
from which the content is to be loaded. A new 'getContent' utility function
is added to simplify its use in components. The existing 'front_page_content'
option is updated to make use of this, meaning that any deployments using that
option will have to be updated with this version of 'shacl-vue'.

This feature will also support, among others, the TTL templating feature of
the 'InstancesUploadEditor'.