Add cancelButtonPressed ref to form
This is set to true whenever a form is cancelled. It is provided and can be injected
by any component wanting to watch for that specific event. This is useful for the
upcoming 'InstancesUploadEditor' in order to know that, when the form is cancelled,
any records created during the upload process should be removed from the store.