'InstancesSelectEditor' updates to improve integration with other components
Firstly, this component was updated to use the new 'useCompConfig' composable
for generic component loading.

Secondly, the 'triggerListGenAndItemSelect' injected ref was introduced.
This ref is watched together with the 'modelValue' prop, and allows triggering
the 'InstancesSelectEditor' to recalculate its list of items and to set the
selected item. Any parent component that initializes and provides this ref
can trigger its 'InstancesSelectEditor' child in that way. This is useful for
the upcoming 'InstancesUploadEditor'.