Allow record pid to be used in display_name_autogenerate functionality
This was not possible because, prior to this commit, the 'getRecordDisplayLabel' function
was only using related quads as inout the the template resolution step, and the pid of
the record itself was not a related quad since it is not a quad onto itself. The pid IS
the named node. Therefore a fauw pid triple had to be added to the input for template
resolution. In addition, recursive resolution had to get a check to endsure that infdefinite
recursion did not happen, for which the id_iri config option was added to configVarsMain