More visible and invasive warning system for tokens and submissions
Introduces the new 'submitWarning' ref, operating on the same principle as the
previously introduced 'tokenWarning' ref.

Now, when a form is saved and there are nodes to submit, the 'submitWarning'
ref is triggered. If there is no token at that point, the 'tokenWarning' ref
is triggered AND the token entry form is automatically opened. If the token
entry form is saved, the 'submitWarning' ref is triggered.

'submitWarning' will let the submitt button flash green for 4 seconds.