Allow component-level config for 'InstancesSelectEditor' 'Add new item' functionality
Up until now, the applicaiton level config options for 'show_classes', 'hide_classes',
'show_classes_with_prefix', and 'hide_classes_with_prefix' have been used in the
'InstancesSelectEditor' to determined which classes to show as menu options when the
user hits the 'Add new item' button. This filtering was necessary because showing all
possible classes by default (which would be any and all subclasses of the class being
rendered by the 'InstancesSelectEditor') could be overwhelming and unnecessary.

However, these app-level config options are not necessarily appropriate for each class
in the context of the 'InstancesSelectEditor'. This can lead, for example, to zero
class options being shown when a user wants to add a new 'Identifier'.

To fix this, the previously introduced 'component_config' is used to allow the same
show/hide config options to be specified per class for the 'InstancesSelectEditor'. An
example config could be:

'component_config': {
    'InstancesSelectEditor': {
        'fetchingsRecordsText': 'Fetching records...',
        'dlidentifiers:Identifier': {
            'show_classes': [],
            'show_classes_with_prefix': ['dlidentifiers'],
            'hide_classes': [],
            'hide_classes_with_prefix': []
        }
    }
}

The 'InstancesSelectEditor' will now use the component- and class-specific config for
show/hide options, and will default to the app-level config for those same options if
the former is not provided.

To support this, the 'includeClass' mehtod was generalized and moved to utils.