Add the 'AssociationClassEditor'
An association class is a class without a PID that has the function of linking records that have
PIDs. For example, a 'Publication' can be 'attributed_to' an 'Agent', with some extra annotations
attached to it, e.g. role, or start/end time, or whatever else. In this case, an 'Agent' (PID
record) can be linked to a 'Publication' (PID record) via an 'AgentAttribution' (association class
record). And that linking record can qualify the relationship further, e.g. the 'role' of 'Author'.

Before the 'AssociationClassEditor', the only way to create this linkage was to explicitly create
the association class record, then select the agent from a list, then qualify the relation more
if necessary, then make sure that the association class record is correctly linked to the original
PID record. This can be both time-consuming and confusing if a user only wants to link an 'Agent'
to a 'Publication', for example. If we can assume that a specific field of the association class
record carries the weight of that record, we can present that field to the user directly, and have
'shacl-vue' handle the association class creation and linkage internally.

That is exactly the purpose of the new 'AssociationClassEditor', with the help of the DASH property
role 'KeyInfoRole', see: https://www.datashapes.org/propertyroles.html\#KeyInfoRole.

The new editor has the following working principles:

- matches when the property shape has 'sh:nodeKind' being either 'sh:IRI' or 'sh:BlankNodeOrIRI'
  AND it has 'sh:class' AND the nodeshape for the target 'sh:class' does not have a PID property
  AND one of these properties is annotated with 'dash:propertyRole': 'dash:KeyInfoRole'.
- it uses '<component :is=>' to dynamically render the component for the property annotated with
  'dash:propertyRole'. This means it will render the correct component for the key field, be it
  an 'InstancesSelectEditor' or 'TextFieldEditor' or whatever.
- when it mounts, it creates the association class record quad, and adds it to the graph, if it
  does not exist yet.
- it can receive component-specific configuration to allow setting default values for related
  properties of an association class record of a specific type; for example, the following
  'component_config' entry tells the 'AssociationClassEditor' for a record of type
  'trr379ra:I18NString' that it should by default set the value of the 'trr379ra:language'
  property to 'en'. This will then create (and add to graph) an additional quad directly after
  the association class record quad.

    'AssociationClassEditor': {
        'trr379ra:I18NString': {
            'defaults': {
                'trr379ra:language': 'en'
            }
        }
    }

- Apart from setting relations by default via config, relations are set in one of two ways:
  - when the user updates the value of the dynamically rendered child component (e.g. selecting
    an instance or entering a text value); here it will create a new quad to relate the association
    class record to the child component value/record. If the child component value/record is
    changed again, the editor removes the existing 'keyPropertyRole' quad and adds a new one.
  - when the user hits the edit button for the association class record, it opens up a new form
    where all the extra relations (and also the key property field) can be edited; this editing
    process occurs in exactly the same way as was done before the 'AssociationClassEditor'.
- It provides an edit button, once the 'keyPropertyRole' field has a value, for the association
  class record to be edited.
- It provides an 'information' icon, which will display a 'BlankNodeViewer' for the association
  class record once any additional relations are set on the record. This viewer will show all
  relations.
- upon saving the form, all the necessary blank node quads have already been added to the graph
- upon canceling the form, the created association class record (and any further relations)
  are to removed from the graph again (aka garbage cleanup). A shortcoming here is that the current
  'FormEditor' state management cannot deal with this challenge correctly in all cases. The reason
  is that the 'FormEditor' remounts everytime the form is viewed, i.e. for the first time it is
  openeded as well as when navigating back to it after another interim form was opened.

To achieve the above, the following updates were necessary:

- The new 'AssociationClassEditor' component
- The 'FormEditor' was updated to provide a new 'saveButtonPressed' ref to all other components,
  so that the 'AssociationClassEditor' can distinguish between the originating user actions when
  unmounting, and act accordingly.
- The 'editInstanceItem' function in 'ShaclVue' was updating to include a new 'addQuadsToForm'
  argument (default 'true') that can be used by the 'AssociationClassEditor' to prevent unnecessary
  addition of quads to the 'formData', since it already added them.
- The component matching code in the 'PropertyShapeEditor' now passes additional arguments to
  matchers, which are necessary for the matching logic used by the 'AssociationClassEditor'
- The 'BlankNodeViewer' and 'NamedNodeViewer' components now use a new 'allowLink' prop (default
  true) to allow the 'AssociationClassEditor' to disable links when displaying relations
- Two new functions in the 'utils' module support the 'AssociationClassEditor' matching and
  mounting/rendering logic: 'getNodeShapePropertyWithAnnotations' and
  'nodeShapeHasPropertyWithAnnotations'

TODO: An existing idea to deal with the 'FormEditor' state issue is to store the state of the opened
form in the 'openForms' global ref, which already exists and is used to know which forms should be
open at any given time, and in which order. This approach could also solve the problem where the
value for the 'All fields' checkbox changes back to false for a form that is navigated back to.
Currently, this is not particularly a show-stopper, because the result of not dealing with this is
just a bunch of dangling quads in the graph store, which will be gone once the session stops.
Dangling quads are also not submitted because they are linked to a non-existing blank node record.

TODO: The way that the 'AssociationClassEditor' reacts to updates from the child component is still
insufficient. These update events are necessary in order to create the 'keyPropertRole' quad on the
fly and add it to the graph: specifically, this needs to happen when there is an update from the
dynamically mounted child component that renders the 'keyPropertyRole' field. This already works,
but it is not consistent across component types. For an 'InstancesSelectEditor' the functionality
is perfect, but for a 'TextFieldEditor' the update happens with each keystroke. The resulting
behaviour is that quads are destroyed and then created anew for each keystroke. This is wasteful,
but still ends up with the correct triples in the end. Further work can look into how such an update
from the child component can be triggered at the end of the entry, rather than for each keystroke.
An alternative would be that quad deletion and creation should not happen 'onUpdate' from the child
component, but rather based on some other event or point in time (e.g. when the user hits the edit
or save button).