Store initial and reactive state of show_all_fields in openForms
There has been a long unadressed issue of the FormEditor's remounting behaviour,
which causes state to be reset upon remounting, even if the formwas already opened
before. This typically happened when a user closes a different form that then brings
the focus back to the current form, which then leads to it remounting instead of
just having maintained the previous state.

This commit does not address that issue fundamentally, but rather introduces a pattern
that could be used to work around the mounting limitations. The pattern involves
storing/tracking state in the provided 'openForms' ref. In the current case, the
'show_all_fields' option is tracked in this way, so that the previous decisions by
the user (to switch it on or off) are maintained across form opening and closing.